buildtools/burg/trim.c(332): remark #2259: non-pointer conversion from "int" to "short" may lose significant bits
http://monetdb.cwi.nl/testing/projects/monetdb/testweb/logs/39245:124139f4ff4a/Int-Fedora-i386-propcheck/make.html#l805

+++ buildtools/burg/trim.c
@@ -329,7 +329,7 @@ trim(t) Item_Set t;
 	}
 	for (m = 1; m < max_nonterminal; m++) {
 		if (t->virgin[m].rule) {
-			vec[last++] = m;
+			vec[last++] = (short)m;
 		}
 	}
 	for (m = 0; m < last; m++) {
